refactor(backend): Round gender ratios to integers in PublicDashboardDto
This commit refactors the `PublicMaleRatio` and `PublicFemaleRatio` properties in `PublicDashboardDto` from `double` to `int`.

The `GetPublicDashboardQueryHandler` has been updated to:
- Calculate gender ratios as percentages.
- Round and cast these percentages to integers before assigning them to the DTO.

The unit tests in `GetPublicDashboardQueryHandlerTests` have been adjusted to assert against these new integer percentage values.

+++ apps/backend/src/Application/Dashboard/Queries/GetPublicDashboard/PublicDashboardDto.cs
@@ -6,8 +6,8 @@ public class PublicDashboardDto
     public int TotalPublicMembers { get; set; }
     public int TotalPublicRelationships { get; set; }
     public int TotalPublicGenerations { get; set; }
-    public double PublicMaleRatio { get; set; }
-    public double PublicFemaleRatio { get; set; }
+    public int PublicMaleRatio { get; set; }
+    public int PublicFemaleRatio { get; set; }
     public int PublicLivingMembersCount { get; set; }
     public int PublicDeceasedMembersCount { get; set; }
     public int PublicAverageAge { get; set; }
